requirements.txt
biopython==1.76
flask
numpy
flask_debug
flask_login